NaturalIsotopeCorrection

Documentation for NaturalIsotopeCorrection.

NaturalIsotopeCorrection._standard_element_MID

NaturalIsotopeCorrection.element_CM

NaturalIsotopeCorrection.fragment_CM

NaturalIsotopeCorrection.isotope_correction

NaturalIsotopeCorrection._standard_element_MID — Method

_standard_element_MID(isotopes, element, N)

Outputs the classic standard mass isotopomer distribution (MID) for a fragment with N atoms of the given element, using combinatorial probabilities. Not to be used for atoms labeled through an isotope tracer experiment, e.g. backbone carbon atoms in a 13C tracer experiment.

Arguments

isotopes::Dict: dictionary containing the natural abundances of the relevant isotopes.

element: the element the standard MID should be computed for.

N::Int: the number atoms of the given element.

NaturalIsotopeCorrection.element_CM — Method

element_CM(element, N)

Outputs the correction matrix for a fragment with N atoms of the given element. All correction matricies will be l+1 x l+1 with l being the total number of possibly labeled atoms.

Arguments

element: the element the standard MID should be computed for.

N::Int: the number atoms of the given element.

l::Int: total number of possibly labeled atoms, default is -1, only needed for unlabeled atoms.

tracer_purity: purity of the tracer, default is 1 (100%).

NaturalIsotopeCorrection.fragment_CM — Method

fragment_CM(formula)

Outputs the full correction matrix for a fragment. The correction matricies will be l+1 x l+1 with l being the total number of possibly labeled atoms.

Arguments

formula: chemical formula of the fragment to be corrected.

tracer_purity: purity of the tracer, default is 1 (100%).

NaturalIsotopeCorrection.isotope_correction — Method

isotope_correction(response, formula)

Outputs the corrected LC-/GCMS response, mass isotopomer distribution (MID) and mean enrichment of a fragment. Additionally outputs the residuum when using L-BFGS-B for optimization.

Arguments

response::Vector: the response / measurement vector obtained from LC-/GCMS, must have the length l+1 if l is equal to the amount of possibly labeled atoms in the fragment

formula::String: chemical formula of the fragment to be corrected.

tracer_purity = 1.0: purity of the tracer, the default represents 100%.

optimization::Bool = true: whether to use the L-BFGS-B optimization algorithm.

Examples

julia> pyruvate = "C3H3O3LabC3"
julia> response_vec = [3500000, 1000000, 3500000, 800000]
julia> corr_response, MID, mean_enrichment, residuum = isotope_correction(response, pyruvate; tracer_purity = 0.99)
([3.6690841286892043e6, 880069.4781954776, 3.548554810172248e6, 765218.8519553166], 
[0.41398107163956405, 0.09929783371601102, 0.4003818041674765, 0.08633929047694855], 
0.5397698283704524, 
[-1.0858374563130465e-13, -2.566428685730154e-15, 1.1059455573558807e-13, -1.6554787924343889e-13])
julia> corr_response, MID, mean_enrichment = isotope_correction(response, pyruvate; tracer_purity = 0.99, optimization = false)
([3.6690841286882004e6, 880069.4781954774, 3.548554810173271e6, 765218.8519538215], 
[0.41398107163951964, 0.09929783371602752, 0.40038180416765856, 0.08633929047679423], 
0.5397698283704317)